ms-ordenes/requirements.txt
Flask==2.1.1
SQLAlchemy==1.4
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0
Flask-JWT-Extended==4.3.1
Werkzeug>=2.0
psycopg2-binary==2.9.3
marshmallow-sqlalchemy==0.28.0
validators==0.18.2
requests==2.27.1
pandas==1.4.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
pytest==7.1.3
Flask-Migrate==3.1.0
pulsar-client==2.10.2
pulsar-client[avro]==2.10.2
